Make sure Frontend elements load the CSS properly
