+++ plugins/common/src/Tribe/Editor.php
@@ -94,93 +94,6 @@ class Tribe__Gutenberg__Common__Editor {
 	}
 
 	/**
-	 * Making sure we have correct post content for blocks after going into Gutenberg
-	 *
-	 * @since  0.2.7-alpha
-	 *
-	 * @param  int $post Which post we will migrate
-	 *
-	 * @return bool
-	 */
-	public function update_post_content_to_blocks( $post ) {
-		$post = get_post( $post );
-		$blocks = $this->get_classic_template();
-		$content = array();
-
-		foreach ( $blocks as $key => $block_param ) {
-			$slug = reset( $block_param );
-
-			if ( 'core/paragraph' === $slug ) {
-				$content[] = '<!-- wp:freeform -->';
-				$content[] = $post->post_content;
-				$content[] = '<!-- /wp:freeform -->';
-			} else {
-				$content[] = '<!-- wp:' . $slug . ' /-->';
-			}
-		}
-
-		$status = wp_update_post( array(
-			'ID' => $post->ID,
-			'post_content' => implode( "\n\r", $content ),
-		) );
-
-		return $status;
-	}
-
-	/**
-	 * When initially loading a post in gutenberg flags if came from classic editor
-	 *
-	 * @since  0.2.7-alpha
-	 *
-	 * @return bool
-	 */
-	public function flag_post_from_classic_editor() {
-		$post = tribe_get_request_var( 'post' );
-
-		// Bail on empty post
-		if ( empty( $post ) ) {
-			return false;
-		}
-
-		// Bail on non numeric post
-		if ( ! is_numeric( $post ) ) {
-			return false;
-		}
-
-		$on_classic_editor_page = tribe_get_request_var( 'classic-editor', false );
-		// Bail if in classic editor
-		if ( $on_classic_editor_page ) {
-			return false;
-		}
-
-		// Bail if not an event (might need to be removed later)
-		if ( ! tribe_is_event( $post ) ) {
-			return false;
-		}
-
-		// Bail if it already has Blocks
-		if ( has_blocks( $post ) ) {
-			return false;
-		}
-
-		$status = update_post_meta( $post, $this->key_flag_classic_editor, 1 );
-
-		// Only trigger when we actually have the correct post
-		if ( $status ) {
-			/**
-			 * Triggers
-			 *
-			 * @since  0.2.2-alpha
-			 *
-			 * @param  int $post Which post is getting updated
-			 */
-			do_action( 'tribe_blocks_editor_flag_post_classic_editor', $post );
-		}
-
-		return $status;
-	}
-
-	/**
 	 * Check if post is from classic editor
 	 *
 	 * @since  0.2.7-alpha
