+++ plugins/events/src/Tribe/Blocks/Classic_Event_Details.php
@@ -30,4 +30,24 @@ extends Tribe__Gutenberg__Events__Blocks__Abstract {
 
 		return tribe( 'gutenberg.events.template' )->template( array( 'blocks', $this->slug() ), $args, false );
 	}
+
+	/**
+	 * Register the Assets for when this block is active
+	 *
+	 * @since  0.2.8-alpha
+	 *
+	 * @return void
+	 */
+	public function assets() {
+		tribe_asset(
+			tribe( 'gutenberg.events.plugin' ),
+			'tribe-events-block-' . $this->slug(),
+			$this->slug() . '/frontend.css',
+			array(),
+			'wp_enqueue_scripts',
+			array(
+				'conditionals' => array( $this, 'has_block' ),
+			)
+		);
+	}
 }
