+++ plugins/events/src/Tribe/Provider.php
@@ -76,33 +76,10 @@ class Tribe__Gutenberg__Events__Provider extends tad_DI52_ServiceProvider {
 	 *
 	 */
 	protected function hook() {
-		add_filter( 'tribe_events_register_event_type_args', tribe_callback( 'gutenberg.events.editor', 'add_event_template_blocks' ) );
-		add_filter( 'tribe_events_register_event_type_args', tribe_callback( 'gutenberg.events.editor', 'add_template_blocks' ) );
-
-		// Add Rest API support
-		add_filter( 'tribe_events_register_event_type_args', tribe_callback( 'gutenberg.events.editor', 'add_rest_support' ) );
-		add_filter( 'tribe_events_register_venue_type_args', tribe_callback( 'gutenberg.events.editor', 'add_rest_support' ) );
-		add_filter( 'tribe_events_register_organizer_type_args', tribe_callback( 'gutenberg.events.editor', 'add_rest_support' ) );
-
-		// Remove assets that are not relevant for Gutenberg Editor
-		add_action( 'wp_print_scripts', tribe_callback( 'gutenberg.events.editor', 'deregister_scripts' ) );
-
 		// Setup the Meta registration
 		add_action( 'init', tribe_callback( 'gutenberg.events.meta', 'register' ), 25 );
 
-		// Setup the registration of Blocks
-		add_action( 'init', tribe_callback( 'gutenberg.events.editor', 'register_blocks' ), 20 );
-		// Load assets of the blocks
-		add_action( 'admin_init', tribe_callback( 'gutenberg.events.editor', 'assets' ) );
-
-		// Maybe add flag from classic editor
-		add_action( 'init', tribe_callback( 'gutenberg.events.editor', 'flag_post_from_classic_editor' ), 0 );
-
-		// Add Block Categories to Editor
-		add_action( 'block_categories', tribe_callback( 'gutenberg.events.editor', 'block_categories' ), 10, 2 );
-
-		// Update Post content to use blocks
-		add_action( 'tribe_blocks_editor_flag_post_classic_editor', tribe_callback( 'gutenberg.events.editor', 'update_post_content_to_blocks' ) );
+		tribe( 'gutenberg.events.editor' )->hook();
 
 		// Register blocks to own own action
 		add_action( 'tribe_events_editor_register_blocks', tribe_callback( 'gutenberg.events.blocks.classic-event-details', 'register' ) );
