--- plugins/events/src/styles/classic-event-details/backend.js
@@ -1,1 +1,0 @@
-import './backend.pcss';
