--- plugins/events/src/styles/classic-event-details/backend.pcss
@@ -1,9 +1,0 @@
-/**
- * Classic Event Details block view styles
- *
- * @since  0.2.4-alpha
- */
-
-.random {
-	color: red;
-}
