--- plugins/events/src/styles/classic-event-details/style.pcss
@@ -1,5 +1,0 @@
-/**
- * Classic Event Details block view styles
- *
- * @since  0.2.4-alpha
- */
