+++ plugins/events/src/styles/event-category/frontend.pcss
