+++ plugins/events/src/styles/event-datetime/frontend.js
@@ -1,0 +1,1 @@
+import './frontend.pcss';
