+++ plugins/events/src/styles/event-links/frontend.pcss
