+++ plugins/events/src/styles/event-organizer/frontend.pcss
