+++ plugins/events/src/styles/event-price/frontend.pcss
