+++ plugins/events/src/styles/event-tags/frontend.pcss
