+++ plugins/events/src/styles/event-venue/frontend.js
@@ -1,0 +1,1 @@
+import './frontend.pcss';
