+++ plugins/events/src/styles/event-website/frontend.js
@@ -1,0 +1,1 @@
+import './frontend.pcss';
