--- plugins/events/src/styles/index.js
@@ -1,13 +1,0 @@
-// General FE styles
-import './style.pcss';
-
-// Import each block view
-import './classic-event-details/style.pcss';
-import './event-datetime/style.pcss';
-import './event-venue/style.pcss';
-import './event-organizer/style.pcss';
-import './event-links/style.pcss';
-import './event-price/style.pcss';
-import './event-category/style.pcss';
-import './event-tags/style.pcss';
-import './event-website/style.pcss';
